finally, a system organisation...
I had initially thought that package inferred systems were the way
forward, but they come with a lot of baggage that I didn't enjoy. I
have reverted to the way I used to do things with a bunch of asdf
systems stacked inside the overarching organisational package, in this
case 'fd-accessors.